+++ hoomd/md/NeighborListGPU.cu
@@ -103,7 +103,7 @@ hipError_t gpu_nlist_needs_update_check_new(unsigned int *d_result,
                                              const unsigned int checkn,
                                              const GPUPartition& gpu_partition)
     {
-    const unsigned int shared_bytes = sizeof(Scalar) * ntypes;
+    const unsigned int shared_bytes = (unsigned int)sizeof(Scalar) * ntypes;
 
     unsigned int block_size = 128;
 
@@ -433,7 +433,7 @@ hipError_t gpu_nlist_build_head_list(unsigned int *d_head_list,
         }
 
     unsigned int run_block_size = min(block_size, max_block_size);
-    unsigned int shared_bytes = ntypes*sizeof(unsigned int);
+    unsigned int shared_bytes = (unsigned int)ntypes*sizeof(unsigned int);
 
     // initialize each particle with its number of neighbors
     hipLaunchKernelGGL((gpu_nlist_init_head_list_kernel), dim3(N/run_block_size + 1), dim3(run_block_size), shared_bytes, 0, d_head_list,
